+++ CHANGELOG.md
@@ -3,19 +3,28 @@
 ## TBD
 * Add media
 * Bulk Edit working
+* Call manage_posts_custom_column and manage_post_posts_columns on demand
 * Correct (de)activate and uninstall routines
 * Correct settings page URL
 * Disable General section for now
+* Display correct field label in editor
+* Don't echo and set columns if not enabled
+* Don't load JavaScript if not needed
 * Dynamically adapt for post_excerpt
 * Enable XSS EscapeOutput checking
 * Initialize localization
+* Link to premium on plugins page
 * Move plugin load to plugins_loaded
 * PHP 5.3 WordPress master testing
+* Pull fields and labels from various custom field helpers
 * Quick Edit working
 * Reduce setting options to core
 * Remove class from settings page link
 * Rename filters and methods to fit plugin naming
+* Static cache enabled fields
 * Test grabbing post types
+* Update API
+* Update POT
 
 ## 0.0.1
 * Initial code release